add(CRUD): add logica nos modulos BOOKING/USERS/SERVICES/AVAILABILITY

+++ src/services/service.service.ts
@@ -1,22 +1,47 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateServiceDto } from './dto/create-service.dto';
 import { UpdateServiceDto } from './dto/update-service.dto';
 import { PrismaService } from 'src/common/prisma/prisma.service';
+import { UsersService } from 'src/users/users.service';
+import { SetRoleAccess } from 'src/auth/decorator/set-role.decorator';
+import { Roles } from 'src/auth/enum/roles.enum';
 
 @Injectable()
 export class ServiceService {
-  constructor(private readonly prismaService: PrismaService) {}
+  constructor(
+    private readonly prismaService: PrismaService,
+    private readonly usersService: UsersService,
+  ) {}
 
-  create(createServiceDto: CreateServiceDto) {
-    return this.prismaService.service.create({
+  async create(createServiceDto: CreateServiceDto) {
+    const service = await this.prismaService.service.create({
       data: {
         ...createServiceDto,
       },
     });
+
+    await this.usersService.update(createServiceDto.providerId, {
+      role: 'PROVIDER',
+    });
+
+    return service;
   }
 
   findAll() {
     return this.prismaService.service.findMany();
+  }
+
+  @SetRoleAccess(Roles.PROVIDER)
+  findAllMyService(id: number) {
+    return this.prismaService.service.findMany({
+      where: {
+        id,
+      },
+    });
   }
 
   findOne(id: number) {
@@ -27,7 +52,23 @@ export class ServiceService {
     });
   }
 
-  update(id: number, updateServiceDto: UpdateServiceDto) {
+  async update(id: number, userId: number, updateServiceDto: UpdateServiceDto) {
+    const service = await this.prismaService.service.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!service) {
+      throw new NotFoundException('Service not found');
+    }
+
+    if (service.providerId !== userId) {
+      throw new ForbiddenException(
+        'You are not allowed to update this service.',
+      );
+    }
+
     return this.prismaService.service.update({
       where: {
         id,
@@ -39,10 +80,53 @@ export class ServiceService {
     });
   }
 
-  remove(id: number) {
+  async remove(id: number, idAuthenticated: number) {
+    const service = await this.prismaService.service.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!service) {
+      throw new NotFoundException('Service not found');
+    }
+
+    if (service.providerId !== idAuthenticated) {
+      throw new ForbiddenException(
+        'You are not allowed to update this service.',
+      );
+    }
+
     return this.prismaService.service.delete({
       where: {
         id,
+      },
+    });
+  }
+
+  async toggleIsActive(id: number, idAuthenticated: number) {
+    const service = await this.prismaService.service.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!service) {
+      throw new NotFoundException('Service not found');
+    }
+
+    if (service.providerId !== idAuthenticated) {
+      throw new ForbiddenException(
+        'You are not allowed to update this service.',
+      );
+    }
+
+    return this.prismaService.service.update({
+      where: {
+        id,
+      },
+      data: {
+        isActive: !service.isActive,
       },
     });
   }
